refactor: Move resolve rule to resolve module
