+++ src/python/pants/backend/javascript/goals/lockfile.py
@@ -13,7 +13,7 @@ from pants.backend.javascript.nodejs_project_environment import (
     NodeJsProjectEnvironmentProcess,
 )
 from pants.backend.javascript.package_json import PackageJsonTarget
-from pants.backend.javascript.subsystems.nodejs import UserChosenNodeJSResolveAliases
+from pants.backend.javascript.resolve import NodeJSProjectResolves, UserChosenNodeJSResolveAliases
 from pants.core.goals.generate_lockfiles import (
     GenerateLockfile,
     GenerateLockfileResult,
@@ -92,24 +92,15 @@ async def determine_package_json_user_resolves(
 
 @rule
 async def setup_user_lockfile_requests(
+    resolves: NodeJSProjectResolves,
     requested: RequestedPackageJsonUserResolveNames,
-    all_projects: AllNodeJSProjects,
-    user_chosen_resolves: UserChosenNodeJSResolveAliases,
 ) -> UserGenerateLockfiles:
-    def get_name(project: NodeJSProject) -> str:
-        return user_chosen_resolves.get(
-            os.path.join(project.root_dir, project.lockfile_name), project.default_resolve_name
-        )
-
-    projects_by_name = {get_name(project): project for project in all_projects}
     return UserGenerateLockfiles(
         GeneratePackageLockJsonFile(
             resolve_name=name,
-            lockfile_dest=os.path.join(
-                projects_by_name[name].root_dir, projects_by_name[name].lockfile_name
-            ),
+            lockfile_dest=os.path.join(resolves[name].root_dir, resolves[name].lockfile_name),
             diff=False,
-            project=projects_by_name[name],
+            project=resolves[name],
         )
         for name in requested
     )
