+++ src/python/pants/backend/javascript/resolve.py
@@ -14,12 +14,15 @@ from pants.backend.javascript.package_json import (
     OwningNodePackageRequest,
     PackageJsonSourceField,
 )
+from pants.backend.javascript.subsystems.nodejs import NodeJS
 from pants.build_graph.address import Address
 from pants.engine.fs import PathGlobs
 from pants.engine.internals.selectors import Get
 from pants.engine.rules import Rule, collect_rules, rule
 from pants.engine.target import Target, WrappedTarget, WrappedTargetRequest
 from pants.engine.unions import UnionRule
+from pants.util.frozendict import FrozenDict
+from pants.util.logging import LogLevel
 
 
 @dataclass(frozen=True)
@@ -66,6 +69,31 @@ async def resolve_for_package(
     directory = await _get_node_package_json_directory(req)
     project = all_projects.project_for_directory(directory)
     return ChosenNodeResolve(project)
+
+
+class NodeJSProjectResolves(FrozenDict[str, NodeJSProject]):
+    pass
+
+
+class UserChosenNodeJSResolveAliases(FrozenDict[str, str]):
+    pass
+
+
+@rule
+async def resolve_to_projects(
+    all_projects: AllNodeJSProjects, user_chosen_resolves: UserChosenNodeJSResolveAliases
+) -> NodeJSProjectResolves:
+    def get_name(project: NodeJSProject) -> str:
+        return user_chosen_resolves.get(
+            os.path.join(project.root_dir, project.lockfile_name), project.default_resolve_name
+        )
+
+    return NodeJSProjectResolves((get_name(project), project) for project in all_projects)
+
+
+@rule(level=LogLevel.DEBUG)
+async def user_chosen_resolve_aliases(nodejs: NodeJS) -> UserChosenNodeJSResolveAliases:
+    return UserChosenNodeJSResolveAliases((value, key) for key, value in nodejs.resolves.items())
 
 
 def rules() -> Iterable[Rule | UnionRule]:
